Refactor: moved common functionality of github/git url installers into base class
